bug fix on resintsllation

+++ nginx/nginx-install.sh
@@ -2370,8 +2370,14 @@ case $OPTION in
         mkdir -p /etc/nginx/conf.d >> /tmp/nginx-install.log 2>&1
         wget -O /etc/nginx/conf.d/geo_fence.conf.default https://raw.githubusercontent.com/marirs/autoinstalls/master/nginx/conf/geo_fence.conf >> /tmp/nginx-install.log 2>&1
 	if [[ "$GEOIP2" = 'y' ]]; then
-	    wget -O /etc/nginx/conf.d/geoip2.conf https://raw.githubusercontent.com/marirs/autoinstalls/master/nginx/conf/geoip2.conf >> /tmp/nginx-install.log 2>&1
-	    wget -O /etc/nginx/conf.d/logformat.conf https://raw.githubusercontent.com/marirs/autoinstalls/master/nginx/conf/logformat.conf >> /tmp/nginx-install.log 2>&1
+	    # Only download geoip2.conf if it doesn't exist to avoid duplication
+	    if [[ ! -e /etc/nginx/conf.d/geoip2.conf ]]; then
+	        wget -O /etc/nginx/conf.d/geoip2.conf https://raw.githubusercontent.com/marirs/autoinstalls/master/nginx/conf/geoip2.conf >> /tmp/nginx-install.log 2>&1
+	    fi
+	    # Only download logformat.conf if it doesn't exist
+	    if [[ ! -e /etc/nginx/conf.d/logformat.conf ]]; then
+	        wget -O /etc/nginx/conf.d/logformat.conf https://raw.githubusercontent.com/marirs/autoinstalls/master/nginx/conf/logformat.conf >> /tmp/nginx-install.log 2>&1
+	    fi
 fi
 		# Nginx installation from source does not add an init script for systemd and logrotate
 		# Using the official systemd script and logrotate conf from nginx.org
